cleaned clippy warnings

+++ sqld/src/replication/primary/logger.rs
@@ -120,7 +120,7 @@ unsafe impl WalHook for ReplicationLoggerHook {
                     replicator.set_page_size(page_size as usize)?;
                     let frame_count = PageHdrIter::new(page_headers, page_size as usize).count();
                     replicator.submit_frames(frame_count as u32);
-                    let last_consistent_frame = last_valid_frame as u32 + frame_count as u32;
+                    let last_consistent_frame = last_valid_frame + frame_count as u32;
                     if is_commit != 0 {
                         replicator.request_flush();
                         replicator
